Leverages FactoryGirl’s Sequences feature to arrive at a unique source_id
https://github.com/thoughtbot/factory_girl/blob/master/GETTING_STARTED.md#sequences

Interestingly enough, using Faker’s random 5 digit option ended up yielding the same number twice.  Highly unlikely, but it did happen.  And this resulted in a validation issue, which then resulted in a test failure.

Trying out FactoryGirl’s sequence feature instead, for the unique key, as this is built into the factory.